+++ api/src/main/java/kernel/maidlab/api/auth/dto/request/SocialLoginRequestDto.java
@@ -1,10 +1,9 @@
 package kernel.maidlab.api.auth.dto.request;
 
-import jakarta.validation.constraints.NotBlank;
-
 import kernel.maidlab.common.enums.SocialType;
 import kernel.maidlab.common.enums.UserType;
 
+import jakarta.validation.constraints.NotBlank;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
