+++ api/src/main/java/kernel/maidlab/api/auth/dto/request/SocialSignUpRequestDto.java
@@ -1,12 +1,10 @@
 package kernel.maidlab.api.auth.dto.request;
 
+import kernel.maidlab.common.enums.Gender;
+
 import java.time.LocalDate;
 
 import jakarta.validation.constraints.NotBlank;
-
-import kernel.maidlab.common.enums.Gender;
-import kernel.maidlab.common.enums.UserType;
-
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
