+++ api/src/main/java/kernel/maidlab/api/auth/service/AuthServiceImpl.java
@@ -8,23 +8,14 @@ import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
 import jakarta.transaction.Transactional;
 
-import kernel.maidlab.api.auth.client.GoogleResourceApi;
-import kernel.maidlab.api.auth.dto.GoogleResourceDto;
-import kernel.maidlab.api.auth.dto.GoogleTokenDto;
-import kernel.maidlab.api.auth.dto.request.ChangePwRequestDto;
-import kernel.maidlab.api.auth.dto.request.LoginRequestDto;
-import kernel.maidlab.api.auth.dto.request.SignUpRequestDto;
-import kernel.maidlab.api.auth.dto.request.SocialLoginRequestDto;
-import kernel.maidlab.api.auth.dto.request.SocialSignUpRequestDto;
-import kernel.maidlab.api.auth.dto.response.LoginResponseDto;
-import kernel.maidlab.api.auth.dto.response.SocialLoginResponseDto;
+import kernel.maidlab.api.auth.social.*;
+import kernel.maidlab.api.auth.dto.request.*;
+import kernel.maidlab.api.auth.dto.response.*;
 import kernel.maidlab.api.auth.entity.Consumer;
 import kernel.maidlab.api.auth.entity.Manager;
-import kernel.maidlab.api.auth.jwt.JwtDto;
-import kernel.maidlab.api.auth.jwt.JwtProperties;
-import kernel.maidlab.api.auth.jwt.JwtProvider;
-import kernel.maidlab.api.auth.repository.ConsumerRepository;
-import kernel.maidlab.api.auth.repository.ManagerRepository;
+import kernel.maidlab.api.auth.jwt.*;
+import kernel.maidlab.api.consumer.repository.ConsumerRepository;
+import kernel.maidlab.api.manager.repository.ManagerRepository;
 import kernel.maidlab.api.exception.BaseException;
 import kernel.maidlab.api.util.PasswordUtil;
 import kernel.maidlab.api.util.CookieUtil;
