+++ api/src/main/java/kernel/maidlab/api/consumer/repository/ConsumerRepository.java
@@ -1,12 +1,13 @@
-package kernel.maidlab.api.auth.repository;
+package kernel.maidlab.api.consumer.repository;
+
+import kernel.maidlab.api.auth.entity.Consumer;
 
 import java.util.Optional;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import kernel.maidlab.api.auth.entity.Consumer;
-
 public interface ConsumerRepository extends JpaRepository<Consumer, Long> {
 	Optional<Consumer> findByUuid(String uuid);
+
 	Optional<Consumer> findByPhoneNumber(String phoneNumber);
 }
