+++ api/src/main/java/kernel/maidlab/api/manager/repository/ManagerRepository.java
@@ -1,13 +1,13 @@
-package kernel.maidlab.api.auth.repository;
+package kernel.maidlab.api.manager.repository;
+
+import kernel.maidlab.api.auth.entity.Manager;
 
 import java.util.Optional;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import kernel.maidlab.api.auth.entity.Manager;
-import kernel.maidlab.api.manager.repository.ManagerRepositoryCustom;
-
 public interface ManagerRepository extends JpaRepository<Manager, Long>, ManagerRepositoryCustom {
 	Optional<Manager> findByUuid(String uuid);
+
 	Optional<Manager> findByPhoneNumber(String phoneNumber);
 }
